fix: resolve Windows build issues and implement recursive dependencies
- Fixed MSVC warnings and added missing Windows libraries (ws2_32, crypt32)
- Implemented recursive dependency resolution in DependencyManager
- Added entry point scanning (package.json) and linking
- Renamed project to pkt in CMakeLists.txt

+++ src/Utils.cpp
@@ -253,7 +253,7 @@ int Utils::fuzzyMatch(const std::string& query, const std::string& target) {
     }
     
     if (queryIdx == lowerQuery.size()) {
-        return (matches * 50) / lowerQuery.size();
+        return static_cast<int>((matches * 50) / lowerQuery.size());
     }
     
     return 0;
